requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cloudinary==1.28.1
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-rest-auth==2.2.2
Django==3.2.10
django-allauth==0.47.0
django-braces==1.15.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-countries==7.2.1
django-decouple==2.1
django-environ==0.8.1
django-oauth-toolkit==1.7.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.13.1
google-auth==2.5.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
jwcrypto==1.0
oauthlib==3.1.1
Pillow==9.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
python-dateutil==1.5
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
wrapt==1.13.3
